first automated process

+++ Jenkinsfile
@@ -13,13 +13,20 @@ pipeline{
     }
     stage('Testing'){
       steps{
-        bat 'python test_app.py'
+        if(env.BRANCH_NAME == 'web_interface'){
+          bat 'python test_app.py'
+        }else if(env.BRANCH_NAME == 'develop'){
+          echo "test of develop branch"
+        }else if(env.BRANCH_NAME == 'release'){
+          input "proceed with deployment to live?"
+        }
       }
     }
     stage('Docker images down'){
       steps{
-        bat 'docker rm -f myflaskapp_c'
-        bat 'docker rmi -f myflaskapp'
+        if(env.BRANCH_NAME != 'master'){
+          bat 'docker rm -f myflaskapp_c'
+          bat 'docker rmi -f myflaskapp'
       }
     }
   }
